Use separate test setup
Unfortunately, it's not easily possible to use ESBuild for the test
setup. There are two reasons for this:
1. We can't use the bundled extension for the tests; the tests depend
on multiple files existing and that the extension requires the files
and modules (e.g. for proxyquire).
2. ESBuild compiles to ES6 modules which have additional restrictions
on what can be modified. Therefore, it's not possible to set certain
exported properties to mock them.

The alternative implemented in this commit is to use a separate test
directory for the tests. This directory is not bundled with the
extension. Instead, this directory is only used for running the tests
and contains the TypeScript compiler output instead of the bundled
extension.