Make createMockApiResponse more convenient
So that we're able to set the status value and also not have not have to
build scanned and skipped repos outside of this method call if we won't
don't need to.

For previous tests, we needed to perform checks on scanned & skipped repos
so we needed to build them outside of this method.

When we re-use this method for the VariantAnalysisMonitor, we will just
need a generic ApiResponse so we can create these repos inside the method.