Revert to v3.0 for p-limit package
Version 4.0 has dropped support for Node 16 / CommonJS so we'd have to
first upgrade to ESM before we could use a newer version of this package.

This would involve updating package.json with:
{
  "type": "module"
}

For now let's use the package that just works. We can think of upgrading as
a separate issue.

See:
- https://github.com/sindresorhus/p-limit/issues/63
- https://gist.github.com/sindresorhus/a39789f98801d908bbc7ff3ecc99d99c
- https://github.com/sindresorhus/p-limit/issues/57