Use p-limit to limit concurrency of downloads
This will allow us to limit results to 3 concurrent downloads at a time.

It does this by waiting for the first batch of downloads to finish before
starting a new one.

At the moment this is limiting concurrent downloads per variant analysis.

We could improve this further by employing a global queue (there is a
corresponding package called p-queue which can do this) so that we limit
concurrent downloads over all variant analyses currently running.

For this we'd need to figure out where to initialize the queue and share
it between commands.