Move download functionality into manager
We introduced batching in the variantAnalysisMonitor but this will only
wait for a download _command_ to execute in the extension, rather than
waiting for the download _request_ to finish.

We'd like to do batching on the download requests.

So instead of triggering multiple commands we're going to trigger just
one and give it all of the repos that have results ready for download.

This will call the variantAnalysisManager which now has a method called
`autoDownloadVariantAnalysisResults`. This will receive all the repos
from the command and call `autoDownloadVariantAnalysisResult` in
batches using the p-limit package we introduced previously.