Move variant analysis tests out of cli-integration directory
We've added some tests to the cli-integration folder which don't require
the CLI to be run successfully.  Let's move them in the no-workspace
folder.

The three files are:
- cli-integration/remote-queries/variant-analysis-manager.test.ts
- cli-integration/remote-queries/variant-analysis-monitor.test.ts
- cli-integration/remote-queries/variant-analysis-results-manager.test

We're also moving the variant-analysis-process.test.ts file into
tests/pure-tests/remote-queries as the processor contains helper methods
rather than being a core class.