Mock CLI and extension context
We're no longer able to stub the extension and its context in the same
way we did for our CLI integration tests.

We do however have helper methods for creating smaller dummy versions
in the `no-workspace` folder.

We've had to extract the `createMockCliServer` method into a separate
file so we can re-use it.

We've also had to redefine `storagePath` since it can no longer be
loaded from `global.helper.ts`. The `global.helper.ts` file only
has methods that execute when we run the CLI integration tests, but
not the `no-workspace` tests.