Stub `commands` in no-workspace tests
Since we no longer use a fully formed extension object for our tests,
it will not have access to the real commands we've defined on it.

This means we'll have to stub the `commands.executeCommand` call.

The side-effect is that we're no longer able to test that our
`autoDownloadVariantAnalysisResult` is really calling the API the
way we expect it to do so we'll have to throw away that test.