Get rid of variantAnalysisManager assertions in monitor tests
Since we've had to stub the `extension.commands` call in order to
move these tests to a different folder, we're no longer calling anything
on the variant analysis manager. This would normally happen by triggering
an extension command.

Let's assert that we call the command insteaed of the manager in
our monitor tests and get rid of the manager setup.