Remove public `get` and `set` methods on variantAnalysisManager
We initially introduced these as public methods in order to make
our our `removeVariantAnalysis` test setup possible.

We can instead replace them:

1. We only need to know the number of variant analyses before and
after we attempt to remove one, so we can replace our `getVariantAnalyses`
method with one that just returns the size.

2. We can replace our `setVariantAnalysis` method with calling
`onVariantAnalysisUpdated` which will update the list of variant analyses
for us. We've had to make `onVariantAnalysisUpdated`, but this is a
much smaller change in terms of privacy (and perhaps harmless?) since
  - we also have `onVariantAnalysisSubmitted` which is already public
  - the `onVariantAnalysisUpdated` method is already triggered from outside
    the manager class, from the variantAnalysisMonitor.