Fix removal tests
We weren't accounting for the fact that local queries only get
removed when they're not in progress.

We were also checking the deleted object's type instead of the
newly selected item's type because ... it's a big file.

In the next PR, I'm going to break this down further.

But I'm running out of steam and I'd like to get these tests in
a passing state.