Use `switch` instead of `if/else if/else if`
After adding some `console.log` statements to each of the if / else blocks
in the `handleRemoveHistoryItem` method, I could see all three cases being
executed, instead of just the block which matches the item type.

I'm not quite sure why, since doing:

```
if (true) {
  console.log('1');
} else if (false) {
  console.log('2');
} else if (false) {
  console.log('3');
}
```

will output only `1`.

But after replacing this with a switch statement, the extra calls don't
happen anymore so we only attempt to remove the item once.