Display pop up message when removing an item from query history
Instead of sending this info the logs, we're opening a pop up to let
the user know that the remote query has been removed from query history.

If the query is still in progress, we also let them know that the GitHub
action will also need to be cancelled (and we provide a link to it).