Add virtualization to interpreted results
This adds react-window and uses it to virtualize the interpreted
results. This is a very basic implementation that tries to guess what
the height of an item will be, but doesn't take into account the actual
text size or line height.

This also doesn't automatically adjust the size of the interpreted
results container to the size of the enclosing container. I'm also not
entirely sure that this can be used to virtualize an expanded list,
unless we also set the results container to have a fixed height.